move factory to the separate file
